refactor, moved functions to compare/match
