+++ e2e/config.js
@@ -1,4 +1,12 @@
 module.exports = {
     APP_PACKAGE: 'com.expensify.chat',
+
+    // The port of the testing server that communicates with the app
     SERVER_PORT: 3000,
+
+    // The amount of times a test should be executed for average performance metrics
+    RUNS: 10,
+
+    // The amount of outliers to remove from a dataset before calculating the average
+    DROP_WORST: 1,
 };
